+++ Example/example_main.tf
@@ -64,4 +64,10 @@ module "tfc_rg_override" {
 
   additional_tags = var.additional_tags
 }
+module "local_rg_naming_values" {
+  # Local
+  source = "../../terraform-azurerm-resourcegroup"
+
+  naming_values = var.naming_values_test
+}